+++ GTiff2Tiles.Benchmarks/Program.cs
@@ -142,7 +142,7 @@ namespace GTiff2Tiles.Benchmarks
             Directory.CreateDirectory(gtiff2TilesOutputDirectoryPath);
             Directory.CreateDirectory(gtiff2TilesTempDirectoryPath);
             Image image = new Image(InputFileInfo);
-            await image.GenerateTilesByCroppingAsync(new DirectoryInfo(gtiff2TilesOutputDirectoryPath), MinZ, MaxZ, threadsCount: ThreadsCount).ConfigureAwait(false);
+            await image.GenerateTilesAsync(new DirectoryInfo(gtiff2TilesOutputDirectoryPath), MinZ, MaxZ, threadsCount: ThreadsCount).ConfigureAwait(false);
             stopwatch.Stop();
             Console.WriteLine("GTiff2Tiles process ended.");
             Console.WriteLine("Time passed:");