+++ GTiff2Tiles.Console/Options.cs
@@ -42,13 +42,6 @@ namespace GTiff2Tiles.Console
         public int MaxZ { get; set; }
 
         /// <summary>
-        /// Algorithm to create tiles. Can be "join" or "crop".
-        /// </summary>
-        [Option('a', "algorithm", Required = true,
-            HelpText = "Algorithm to create tiles. Can be \"join\" or \"crop\".")]
-        public string Algorithm { get; set; }
-
-        /// <summary>
         /// Do you want to create tms-compatible tiles?
         /// </summary>
         [Option("tms", Required = true, HelpText = "Do you want to create tms-compatible tiles?")]