+++ GTiff2Tiles.Tests/Tests/Image/ImageTests.cs
@@ -16,59 +16,7 @@ namespace GTiff2Tiles.Tests.Tests.Image
         public void Setup() { }
 
         [Test]
-        public async Task GenerateTmsTilesByJoiningAsync()
-        {
-            DirectoryInfo examplesDirectoryInfo = Helpers.TestHelper.GetExamplesDirectoryInfo();
-
-            DirectoryInfo tempDirectoryInfo =
-                new DirectoryInfo(Path.Combine(Helpers.TestHelper.GetExamplesDirectoryInfo().FullName,
-                                               Enums.FileSystemEntries.Temp,
-                                               DateTime.Now.ToString(DateTimePatterns.LongWithMs)));
-
-            string inputFilePath = Path.Combine(examplesDirectoryInfo.FullName,
-                                                Enums.FileSystemEntries.InputDirectoryName,
-                                                $"{Enums.FileSystemEntries.Input4326}{Extensions.Tif}");
-            string outputDirectoryName = Path.Combine(examplesDirectoryInfo.FullName,
-                                                      Enums.FileSystemEntries.TmsCompatible,
-                                                      Enums.FileSystemEntries
-                                                           .GenerateTilesByJoiningOutputDirectoryName);
-
-            FileInfo inputFileInfo = new FileInfo(inputFilePath);
-            DirectoryInfo outputDirectoryInfo = new DirectoryInfo(outputDirectoryName);
-            IProgress<double> progress = new Progress<double>(Console.WriteLine);
-            const bool tmsCompatible = true;
-
-            try
-            {
-                //Check for errors.
-                CheckHelper.CheckDirectory(outputDirectoryInfo, true);
-
-                if (!await CheckHelper.CheckInputFileAsync(inputFileInfo))
-                {
-                    string tempFilePath = Path.Combine(tempDirectoryInfo.FullName,
-                                                       $"{Gdal.TempFileName}{Extensions.Tif}");
-                    FileInfo tempFileInfo = new FileInfo(tempFilePath);
-
-                    await Core.Image.Gdal.WarpAsync(inputFileInfo, tempFileInfo, Gdal.RepairTifOptions);
-                    inputFileInfo = tempFileInfo;
-                }
-
-                //Create Image object and crop tiles.
-                Core.Image.Image image = new Core.Image.Image(inputFileInfo);
-                const int minZ = Enums.Zooms.MinZ;
-                const int maxZ = Enums.Zooms.MaxZ;
-                const int threadsCount = Enums.Multithreading.ThreadsCount;
-                const string tileExtension = Extensions.Png;
-                await image.GenerateTilesByJoiningAsync(outputDirectoryInfo, minZ, maxZ, tmsCompatible, tileExtension, progress,
-                                                        threadsCount);
-            }
-            catch (Exception exception) { Assert.Fail(exception.Message); }
-
-            Assert.Pass();
-        }
-
-        [Test]
-        public async Task GenerateTmsTilesByCroppingAsync()
+        public async Task GenerateTmsTilesAsync()
         {
             DirectoryInfo examplesDirectoryInfo = Helpers.TestHelper.GetExamplesDirectoryInfo();
 
@@ -111,7 +59,7 @@ namespace GTiff2Tiles.Tests.Tests.Image
                 const int maxZ = Enums.Zooms.MaxZ;
                 const int threadsCount = Enums.Multithreading.ThreadsCount;
                 const string tileExtension = Extensions.Png;
-                await image.GenerateTilesByCroppingAsync(outputDirectoryInfo, minZ, maxZ, tmsCompatible, tileExtension, progress,
+                await image.GenerateTilesAsync(outputDirectoryInfo, minZ, maxZ, tmsCompatible, tileExtension, progress,
                                                     threadsCount);
             }
             catch (Exception exception) { Assert.Fail(exception.Message); }
@@ -120,59 +68,7 @@ namespace GTiff2Tiles.Tests.Tests.Image
         }
 
         [Test]
-        public async Task GenerateNonTmsTilesByJoiningAsync()
-        {
-            DirectoryInfo examplesDirectoryInfo = Helpers.TestHelper.GetExamplesDirectoryInfo();
-
-            DirectoryInfo tempDirectoryInfo =
-                new DirectoryInfo(Path.Combine(Helpers.TestHelper.GetExamplesDirectoryInfo().FullName,
-                                               Enums.FileSystemEntries.Temp,
-                                               DateTime.Now.ToString(DateTimePatterns.LongWithMs)));
-
-            string inputFilePath = Path.Combine(examplesDirectoryInfo.FullName,
-                                                Enums.FileSystemEntries.InputDirectoryName,
-                                                $"{Enums.FileSystemEntries.Input4326}{Extensions.Tif}");
-            string outputDirectoryName = Path.Combine(examplesDirectoryInfo.FullName,
-                                                      Enums.FileSystemEntries.NonTmsCompatible,
-                                                      Enums.FileSystemEntries
-                                                           .GenerateTilesByJoiningOutputDirectoryName);
-
-            FileInfo inputFileInfo = new FileInfo(inputFilePath);
-            DirectoryInfo outputDirectoryInfo = new DirectoryInfo(outputDirectoryName);
-            IProgress<double> progress = new Progress<double>(Console.WriteLine);
-            const bool tmsCompatible = false;
-
-            try
-            {
-                //Check for errors.
-                CheckHelper.CheckDirectory(outputDirectoryInfo, true);
-
-                if (!await CheckHelper.CheckInputFileAsync(inputFileInfo))
-                {
-                    string tempFilePath = Path.Combine(tempDirectoryInfo.FullName,
-                                                       $"{Gdal.TempFileName}{Extensions.Tif}");
-                    FileInfo tempFileInfo = new FileInfo(tempFilePath);
-
-                    await Core.Image.Gdal.WarpAsync(inputFileInfo, tempFileInfo, Gdal.RepairTifOptions);
-                    inputFileInfo = tempFileInfo;
-                }
-
-                //Create Image object and crop tiles.
-                Core.Image.Image image = new Core.Image.Image(inputFileInfo);
-                const int minZ = Enums.Zooms.MinZ;
-                const int maxZ = Enums.Zooms.MaxZ;
-                const int threadsCount = Enums.Multithreading.ThreadsCount;
-                const string tileExtension = Extensions.Png;
-                await image.GenerateTilesByJoiningAsync(outputDirectoryInfo, minZ, maxZ, tmsCompatible, tileExtension, progress,
-                                                        threadsCount);
-            }
-            catch (Exception exception) { Assert.Fail(exception.Message); }
-
-            Assert.Pass();
-        }
-
-        [Test]
-        public async Task GenerateNonTmsTilesByCroppingAsync()
+        public async Task GenerateNonTmsTilesAsync()
         {
             DirectoryInfo examplesDirectoryInfo = Helpers.TestHelper.GetExamplesDirectoryInfo();
 
@@ -215,7 +111,7 @@ namespace GTiff2Tiles.Tests.Tests.Image
                 const int maxZ = Enums.Zooms.MaxZ;
                 const int threadsCount = Enums.Multithreading.ThreadsCount;
                 const string tileExtension = Extensions.Png;
-                await image.GenerateTilesByCroppingAsync(outputDirectoryInfo, minZ, maxZ, tmsCompatible, tileExtension, progress,
+                await image.GenerateTilesAsync(outputDirectoryInfo, minZ, maxZ, tmsCompatible, tileExtension, progress,
                                                          threadsCount);
             }
             catch (Exception exception) { Assert.Fail(exception.Message); }